Reorganised algorithms.gradients (#9969)
* Reorganised algorithms.gradients

Fixes issue #9695 by reorganising algorithms.gradients into seperate folders.

* Moved 'lin_comb_qgt.py' to the 'qgt' folder

Moved 'lin_comb_qgt.py' to the 'qgt' folder to ensure all qgt related files are grouped together.

* Formatted files with linter.

Ran linter and reformatted files accordingly.

* Updated '__init__.py'

Updated '__init__.py' to reflect new folder structure.

* Removed 'qgt' folder

Assimilated QGT classes into related folders within 'algorithms.gradients'. Also, 'qgt_result.py' was placed into the 'base' folder since it constitutes a base class.

* Update qiskit/algorithms/gradients/__init__.py


* Update qiskit/algorithms/gradients/__init__.py


* Removed 'qfi' folder

Removed 'qfi' folder and moved remaining contents into parent folder 'gradients'.

* Update qiskit/algorithms/gradients/__init__.py


* Attempted to resolve merge conflicts.

Altered files to match new versions causing the conflict.

* Ran linter

Ran 'tox -eblack' on the repository to resolve linter issues.

* Fixed base module hierarchy issue

Altered contents of 'qfi.py' to refer to 'BaseQGT' according to the new hierarchy given in the PR.

* Fixed more hierarchy issues.

Changed more references to reflect new hierarchy.

---------
