+++ qiskit/algorithms/gradients/base/estimator_gradient_result.py
