+++ qiskit/algorithms/gradients/base/sampler_gradient_result.py
