+++ qiskit/algorithms/gradients/finite_diff/finite_diff_sampler_gradient.py
@@ -23,10 +23,10 @@ from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.primitives import BaseSampler
 from qiskit.providers import Options
 
-from .base_sampler_gradient import BaseSamplerGradient
-from .sampler_gradient_result import SamplerGradientResult
+from ..base.base_sampler_gradient import BaseSamplerGradient
+from ..base.sampler_gradient_result import SamplerGradientResult
 
-from ..exceptions import AlgorithmError
+from ...exceptions import AlgorithmError
 
 
 class FiniteDiffSamplerGradient(BaseSamplerGradient):
