+++ qiskit/algorithms/gradients/lin_comb/lin_comb_qgt.py
@@ -25,12 +25,12 @@ from qiskit.primitives.utils import _circuit_key
 from qiskit.providers import Options
 from qiskit.quantum_info import SparsePauliOp
 
-from .base_qgt import BaseQGT
+from ..base.base_qgt import BaseQGT
 from .lin_comb_estimator_gradient import LinCombEstimatorGradient
-from .qgt_result import QGTResult
-from .utils import DerivativeType, _make_lin_comb_qgt_circuit, _make_lin_comb_observables
+from ..base.qgt_result import QGTResult
+from ..utils import DerivativeType, _make_lin_comb_qgt_circuit, _make_lin_comb_observables
 
-from ..exceptions import AlgorithmError
+from ...exceptions import AlgorithmError
 
 
 class LinCombQGT(BaseQGT):
