+++ qiskit/algorithms/gradients/param_shift/param_shift_estimator_gradient.py
@@ -21,11 +21,11 @@ from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.opflow import PauliSumOp
 from qiskit.quantum_info.operators.base_operator import BaseOperator
 
-from .base_estimator_gradient import BaseEstimatorGradient
-from .estimator_gradient_result import EstimatorGradientResult
-from .utils import _make_param_shift_parameter_values
+from ..base.base_estimator_gradient import BaseEstimatorGradient
+from ..base.estimator_gradient_result import EstimatorGradientResult
+from ..utils import _make_param_shift_parameter_values
 
-from ..exceptions import AlgorithmError
+from ...exceptions import AlgorithmError
 
 
 class ParamShiftEstimatorGradient(BaseEstimatorGradient):
